+++ src/reporter/reporters/BaseReporter.ts
@@ -1,14 +1,8 @@
-/* eslint-disable no-console */
-import fs from "fs";
 import chalk from "chalk";
 import CliTable3 from "cli-table3";
 
-import { VerifierLanguageType } from "@solarity/zkit";
-
 import { SpinnerProcessor } from "../SpinnerProcessor";
 import { ProgressBarProcessor } from "../ProgressBarProcessor";
-import { HardhatZKitError } from "../../errors";
-import { BYTES_IN_MB } from "../../constants";
 
 export abstract class BaseReporter {
   protected _spinnerProcessor: SpinnerProcessor;
@@ -47,27 +41,6 @@ export abstract class BaseReporter {
 
   protected _getSpinnerWorkingTimeMessage(workingTime: string | undefined): string {
     return workingTime ? chalk.grey(`(${workingTime} s)`) : "";
-  }
-
-  protected _getVerifierLanguageMessage(verifiersType: VerifierLanguageType): string {
-    switch (verifiersType) {
-      case "sol":
-        return "Solidity";
-      case "vy":
-        return "Vyper";
-      default:
-        throw new HardhatZKitError(`Invalid verifiers type - ${verifiersType}`);
-    }
-  }
-
-  protected _getFileSizeInMB(filePath: string | undefined): string {
-    if (!filePath) {
-      throw new HardhatZKitError("File path is undefined. Unable to get file size.");
-    }
-
-    const fileSize: number = fs.statSync(filePath).size;
-
-    return (fileSize / BYTES_IN_MB).toFixed(3);
   }
 
   protected _startSpinner(spinnerIdName: string, spinnerIdSuffix: string, spinnerText: string): string | null {
