+++ src/reporter/reporters/CircuitCompilationReporter.ts
