+++ src/reporter/reporters/ProgressReporter.ts
@@ -1,6 +1,5 @@
 /* eslint-disable no-console */
 import chalk from "chalk";
-import debug from "debug";
 
 import { BaseReporter } from "./BaseReporter";
 
@@ -27,9 +26,5 @@ export class ProgressReporter extends BaseReporter {
 
   public updateProgressBarValue(valueToAdd: number) {
     this._progressBarProcessor.updateProgressBar(valueToAdd);
-  }
-
-  public verboseLog(namespace: string, formatterStr: string, logArgs: any[] = []) {
-    debug(`hardhat-zkit:${namespace}`)(formatterStr, ...logArgs);
   }
 }
